requirements.txt
APScheduler==3.9.1
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-APScheduler==1.12.4
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-WTF==1.0.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
pymongo==4.3.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.9.0
